Rename Yaml Loader to more generic File Loader
