+++ test_runner/src/main/kotlin/ftl/args/AndroidArgs.kt
@@ -25,7 +25,7 @@ import ftl.args.yml.GcloudYml
 import ftl.args.yml.AppTestPair
 import ftl.args.yml.AndroidGcloudYmlParams
 import ftl.args.yml.YamlDeprecated
-import ftl.args.yml.loader.loadYamlFile
+import ftl.util.loadFile
 import ftl.cli.firebase.test.android.AndroidRunCommand
 import ftl.config.Device
 import ftl.config.FtlConstants
@@ -197,7 +197,7 @@ AndroidArgs
             mergeYmlMaps(GcloudYml, AndroidGcloudYml, FlankYml, AndroidFlankYml)
         }
 
-        fun load(yamlPath: Path, cli: AndroidRunCommand? = null): AndroidArgs = load(loadYamlFile(yamlPath), cli)
+        fun load(yamlPath: Path, cli: AndroidRunCommand? = null): AndroidArgs = load(loadFile(yamlPath), cli)
 
         @VisibleForTesting
         internal fun load(yamlReader: Reader, cli: AndroidRunCommand? = null): AndroidArgs {
