+++ test_runner/src/main/kotlin/ftl/util/FileLoader.kt
@@ -1,13 +1,12 @@
-package ftl.args.yml.loader
+package ftl.util
 
-import ftl.util.FlankNoSuchFileException
 import java.io.Reader
 import java.nio.file.Files
 import java.nio.file.NoSuchFileException
 import java.nio.file.Path
 
-fun loadYamlFile(filePath: Path): Reader = try {
+fun loadFile(filePath: Path): Reader = try {
     Files.newBufferedReader(filePath)
 } catch (fileNotFound: NoSuchFileException) {
-    throw FlankNoSuchFileException("YAML file not found: ${fileNotFound.message}")
+    throw FlankNoSuchFileException("File not found: ${fileNotFound.message}")
 }
