+++ test_runner/src/test/kotlin/ftl/args/yml/FileLoaderTest.kt
@@ -1,6 +1,6 @@
 package ftl.args.yml
 
-import ftl.args.yml.loader.loadYamlFile
+import ftl.util.loadFile
 import ftl.test.util.TestHelper.getThrowable
 import ftl.util.FlankNoSuchFileException
 import org.junit.Assert
@@ -8,19 +8,19 @@ import org.junit.Test
 import java.nio.file.Paths
 import java.util.UUID
 
-class YamlLoaderTest {
+class FileLoaderTest {
     @Test(expected = FlankNoSuchFileException::class)
     fun `should throws FlankConfigurationNotFoundException when file not found`() {
         val filePath = Paths.get("${UUID.randomUUID()}.yml")
-        loadYamlFile(filePath)
+        loadFile(filePath)
     }
 
     @Test
     fun `should throws FlankConfigurationNotFoundException with specific message when file not found`() {
         val filePath = Paths.get("${UUID.randomUUID()}.yml")
-        val thrownException = getThrowable { loadYamlFile(filePath) }
+        val thrownException = getThrowable { loadFile(filePath) }
 
-        val expectedExceptionMessage = "YAML file not found: $filePath"
+        val expectedExceptionMessage = "File not found: $filePath"
         Assert.assertEquals(expectedExceptionMessage, thrownException.message)
 
         val expectedExceptionType = FlankNoSuchFileException::class
